Fix --rebind-domain-ok domain matching
Commit 32e15c3f458c2e8838a9ecf7d478ecb6750516bf simplified the
implementation of the rebind-domain-ok feature, but unfortunately, it
also changed its semantics.

In v2.85, --stop-dns-rebind --rebind-domain-ok=test.me would only allow
"test.me" and subdomains of "test.me" to return private addresses to
the user. A query for localtest.me, which is known to return 127.0.0.1,
is blocked as expected.

In v2.86, the --rebind-domain-ok feature is implemented as a
simple suffix comparison, which means that --stop-dns-rebind
--rebind-domain-ok=test.me fails to block the response of "127.0.0.1"
for "localtest.me" because "test.me" is a suffix of "localtest.me".

The special meaning of the empty string is also affected by this bug. In
v2.85, --stop-dns-rebind --rebind-domain-ok=// was interpreted to mean
"stop potential DNS rebinding attacks, but allow responses for dotless
domains". This mirrored the special meaning of // in the --server
option.

In contrast, v2.86, --stop-dns-rebind --rebind-domain-ok=// effectively
negates --stop-dns-rebind because the empty string is a suffix of all
possible strings.

This patch restores the semantics of the --rebind-domain-ok option back
to v2.85.